Implemented stripe payment on the contact checkout page

+++ app/Http/Controllers/View/Auth/AuthViewController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\View\Auth;
 
 use App\Http\Controllers\Controller;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -18,12 +19,57 @@ class AuthViewController extends Controller
     }
 
     public function twoFactor (){
+        if (!Auth::user()->two_factor_code){
+            // Check user roles before redirecting
+            if (Auth::user()->hasRole('super-admin')){
+                return redirect()->route('admin.dashboard');
+            }
+
+            if (Auth::user()->hasRole('company')){
+                return redirect()->route('company.dashboard');
+            }
+        }
         $data = [
             'title' => 'two-factor',
             'keywords' => 'two-factor',
             'description' => 'two-factor'
         ];
         return view('livewire.auth.pages.two-factor-page', ['data' => $data]);
+    }
+
+    public function twoFactorThroughLink ($code){
+        $data = [
+            'title' => 'two-factor',
+            'keywords' => 'two-factor',
+            'description' => 'two-factor'
+        ];
+
+        $user = User::find(Auth::user()->id);
+
+
+        // Check if the code exists
+        if ($code != $user->two_factor_code){
+            // return error message
+            return redirect(route('verify-two-factor'))->with('tokenError', 'Invalid token');
+        }
+
+        // Check if the code is still valid
+        if (now() >= $user->two_factor_expires_at){
+            // return error message
+            return redirect(route('verify-two-factor'))->with('tokenError', 'Two factor expired, please login again');
+        }
+
+        // Delete the old code
+        $user->deleteTwoFactorCode();
+
+        // Check user roles before redirecting
+        if (Auth::user()->hasRole('super-admin')){
+           return redirect()->intended(route('admin.dashboard'));
+        }
+
+        if (Auth::user()->hasRole('company')){
+            return redirect()->intended(route('company.dashboard'));
+        }
     }
 
 
@@ -85,8 +131,10 @@ class AuthViewController extends Controller
 
     public function logout() {
         // Delete the old code
-        Auth::user()->deleteTwoFactorCode();
-        Auth::logout();
+        if (Auth::user()){
+            Auth::user()->deleteTwoFactorCode();
+            Auth::logout();
+        }
         return redirect()->route('sign-in');
     }
 }
